--- app/src/main/java/ru/pashaginas/myapplication/MoneyItem.kt
@@ -1,5 +1,0 @@
-package ru.pashaginas.myapplication
-
-data class MoneyItem(val amount: Int, val purpose: String)
-
-
