+++ app/src/main/java/ru/pashaginas/myapplication/MoneyItemDataClass.kt
@@ -1,0 +1,5 @@
+package ru.pashaginas.myapplication
+
+data class MoneyItemDataClass(val amount: Int, val purpose: String)
+
+
