+++ app/src/main/java/ru/pashaginas/myapplication/adapters/ViewPagerAdapter.kt
@@ -1,21 +1,18 @@
-package ru.pashaginas.myapplication
+package ru.pashaginas.myapplication.adapters
 
 import android.content.res.Resources
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
+import ru.pashaginas.myapplication.fragments.FragmentA
+import ru.pashaginas.myapplication.fragments.FragmentB
+import ru.pashaginas.myapplication.fragments.FragmentC
 
 class ViewPagerAdapter(fragmentActivity: FragmentActivity) :
     FragmentStateAdapter(fragmentActivity) {
     override fun getItemCount(): Int = 3
 
     override fun createFragment(position: Int): Fragment {
-//        if (position == 0) {
-//            return FragmentA()
-//        } else if (position == 1) {
-//            return FragmentB()
-//        }
-//        return FragmentA()
 
         return when (position) {
             0 -> FragmentA()
