+++ app/src/main/java/ru/pashaginas/myapplication/fragments/FragmentA.kt
@@ -1,21 +1,26 @@
-package ru.pashaginas.myapplication
+package ru.pashaginas.myapplication.fragments
 
 import android.os.Bundle
+import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.fragment.app.Fragment
+import ru.pashaginas.myapplication.R
 
-class FragmentC : Fragment() {
+
+//BudgetFragment
+class FragmentA : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragment_c, container, false)
+        return inflater.inflate(R.layout.fragment_a, container, false)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        arguments
     }
 }
