+++ app/src/main/java/ru/pashaginas/myapplication/fragments/FragmentB.kt
@@ -1,19 +1,20 @@
-package ru.pashaginas.myapplication
+package ru.pashaginas.myapplication.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import ru.pashaginas.myapplication.R
 
-class FragmentA : Fragment() {
+class FragmentB : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragment_a, container, false)
+        return inflater.inflate(R.layout.fragment_b, container, false)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
