+++ app/src/main/res/layout/activity_login.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/backgroundColor"
     android:onClick="onClick"
-    tools:context=".LoginActivity">
+    tools:context=".activities.LoginActivity">
 
 
     <ImageView
